+++ pysys-examples/internal/testcases/PySys_internal_068/Input/NestedPass/run.py
@@ -5,5 +5,6 @@ from pysys.exceptions import *
 class PySysTest(BaseTest):
 	def execute(self):
 		self.addOutcome(PASSED)
+		self.reportPerformanceResult(123, 'A performance key', '/s')
 	def validate(self):
 		pass